# Change script and transaction locations to arrays