--- a/runtime/common/scriptlocation.go
+++ b/runtime/common/scriptlocation.go
@@ -29,11 +29,14 @@ const ScriptLocationPrefix = "s"
 
 // ScriptLocation
 //
-type ScriptLocation []byte
+type ScriptLocation [32]byte
 
-func NewScriptLocation(gauge MemoryGauge, script []byte) ScriptLocation {
-	UseMemory(gauge, NewBytesMemoryUsage(len(script)))
-	return ScriptLocation(script)
+var _ Location = ScriptLocation{}
+
+func NewScriptLocation(gauge MemoryGauge, identifier []byte) (location ScriptLocation) {
+	UseMemory(gauge, NewBytesMemoryUsage(len(identifier)))
+	copy(location[:], identifier)
+	return
 }
 
 func (l ScriptLocation) ID() LocationID {
@@ -68,7 +71,7 @@ func (l ScriptLocation) QualifiedIdentifier(typeID TypeID) string {
 }
 
 func (l ScriptLocation) String() string {
-	return hex.EncodeToString(l)
+	return hex.EncodeToString(l[:])
 }
 
 func (l ScriptLocation) MarshalJSON() ([]byte, error) {
@@ -95,7 +98,7 @@ func decodeScriptLocationTypeID(gauge MemoryGauge, typeID string) (ScriptLocatio
 	const errorMessagePrefix = "invalid script location type ID"
 
 	newError := func(message string) (ScriptLocation, string, error) {
-		return nil, "", fmt.Errorf("%s: %s", errorMessagePrefix, message)
+		return ScriptLocation{}, "", fmt.Errorf("%s: %s", errorMessagePrefix, message)
 	}
 
 	if typeID == "" {
@@ -115,7 +118,7 @@ func decodeScriptLocationTypeID(gauge MemoryGauge, typeID string) (ScriptLocatio
 	prefix := parts[0]
 
 	if prefix != ScriptLocationPrefix {
-		return nil, "", fmt.Errorf(
+		return ScriptLocation{}, "", fmt.Errorf(
 			"%s: invalid prefix: expected %q, got %q",
 			errorMessagePrefix,
 			ScriptLocationPrefix,
@@ -127,7 +130,7 @@ func decodeScriptLocationTypeID(gauge MemoryGauge, typeID string) (ScriptLocatio
 	UseMemory(gauge, NewBytesMemoryUsage(len(location)))
 
 	if err != nil {
-		return nil, "", fmt.Errorf(
+		return ScriptLocation{}, "", fmt.Errorf(
 			"%s: invalid location: %w",
 			errorMessagePrefix,
 			err,
@@ -136,5 +139,8 @@ func decodeScriptLocationTypeID(gauge MemoryGauge, typeID string) (ScriptLocatio
 
 	qualifiedIdentifier := parts[2]
 
-	return location, qualifiedIdentifier, nil
+	var result ScriptLocation
+	copy(result[:], location)
+
+	return result, qualifiedIdentifier, nil
 }