--- a/runtime/common/stringlocation.go
+++ b/runtime/common/stringlocation.go
@@ -30,6 +30,8 @@ const StringLocationPrefix = "S"
 //
 type StringLocation string
 
+var _ Location = StringLocation("")
+
 func NewStringLocation(gauge MemoryGauge, id string) StringLocation {
 	UseMemory(gauge, NewRawStringMemoryUsage(len(id)))
 	return StringLocation(id)