--- a/runtime/error_test.go
+++ b/runtime/error_test.go
@@ -60,7 +60,7 @@ func TestRuntimeError(t *testing.T) {
 			err,
 			"Execution failed:\n"+
 				"error: unexpected token: identifier\n"+
-				" --> 01:1:0\n"+
+				" --> 0100000000000000000000000000000000000000000000000000000000000000:1:0\n"+
 				"  |\n"+
 				"1 | X\n"+
 				"  | ^\n",
@@ -93,7 +93,7 @@ func TestRuntimeError(t *testing.T) {
 			err,
 			"Execution failed:\n"+
 				"error: missing access modifier for function\n"+
-				" --> 01:1:0\n"+
+				" --> 0100000000000000000000000000000000000000000000000000000000000000:1:0\n"+
 				"  |\n"+
 				"1 | fun test() {}\n"+
 				"  | ^\n",
@@ -133,7 +133,7 @@ func TestRuntimeError(t *testing.T) {
 			err,
 			"Execution failed:\n"+
 				"error: overflow\n"+
-				" --> 01:6:16\n"+
+				" --> 0100000000000000000000000000000000000000000000000000000000000000:6:16\n"+
 				"  |\n"+
 				"6 |                 a + b\n"+
 				"  |                 ^^^^^\n",
@@ -300,7 +300,8 @@ func TestRuntimeError(t *testing.T) {
 		id, err := hex.DecodeString("57717cc72f97494ac90441790352a07b999a39526819e638b5d367e62e43c37a")
 		require.NoError(t, err)
 
-		location := common.TransactionLocation(id)
+		var location common.TransactionLocation
+		copy(location[:], id)
 
 		codes := map[common.LocationID]string{
 			location.ID(): `