--- a/runtime/sema/check_create_expression.go
+++ b/runtime/sema/check_create_expression.go
@@ -72,7 +72,7 @@ func (checker *Checker) checkResourceCreationOrDestruction(compositeType *Compos
 	contractType := containingContractKindedType(compositeType)
 
 	if contractType == nil {
-		if common.LocationsMatch(compositeType.Location, checker.Location) {
+		if compositeType.Location == checker.Location {
 			return
 		}
 	} else {