Closes #???

## Description

Scripts and transactions are always identified by a 256-bit identifier. 

Make the location types 32-byte arrays. With this change all locations can be used directly as `map` keys, locations can be compared, and a conversion to a location ID is not necessary.

This or a future PR should refactor other cases where location IDs are used as keys for indices (`map`) and use locations directly.
______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels