@turbolent 2022-06-20
Benchmarked this locally, shows a good performance improvement:

```
name                             old time/op    new time/op    delta
RuntimeFungibleTokenTransfer-12     758µs ±17%     651µs ± 9%  -14.16%  (p=0.000 n=20+19)
```